registerTeacher to registerStudent files
